Refactor connector to remove extra methods
Call handlers directly from models
